Extract runtime health hooks and components using AI
