+++ src/components/notebook/NotebookViewer.tsx
@@ -1,6 +1,6 @@
 import { queryDb } from "@livestore/livestore";
 import { useStore } from "@livestore/react";
-import { CellData, events, RuntimeSessionData, tables } from "@runt/schema";
+import { CellData, events, tables } from "@runt/schema";
 import React, { Suspense, useCallback } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
@@ -12,25 +12,25 @@ import { Button } from "@/components/ui/button";
 
 import { useCurrentUserId } from "@/hooks/useCurrentUser.js";
 import { useUserRegistry } from "@/hooks/useUserRegistry.js";
-import { getRuntimeCommand } from "@/util/runtime-command.js";
+import { useRuntimeHealth } from "@/hooks/useRuntimeHealth.js";
+
 import { getCurrentNotebookId } from "@/util/store-id.js";
 import { getClientTypeInfo, getClientColor } from "@/services/userTypes.js";
 import {
   Bot,
   Bug,
   BugOff,
-  Circle,
   Code,
-  Copy,
   Database,
   FileText,
   Filter,
-  Square,
   Terminal,
   X,
 } from "lucide-react";
 import { UserProfile } from "../auth/UserProfile.js";
 import { useAvailableAiModels, getDefaultAiModel } from "@/util/ai-models.js";
+import { RuntimeHealthIndicator } from "./RuntimeHealthIndicator.js";
+import { RuntimeHelper } from "./RuntimeHelper.js";
 
 // Lazy import DebugPanel only in development
 const LazyDebugPanel = React.lazy(() =>
@@ -58,6 +58,7 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
   const currentUserId = useCurrentUserId();
   const { presentUsers, getUserInfo, getUserColor } = useUserRegistry();
   const { models } = useAvailableAiModels();
+  const { runtimeHealth } = useRuntimeHealth();
 
   const cells = store.useQuery(
     queryDb(tables.cells.select().orderBy("position", "asc"))
@@ -93,83 +94,10 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
     queryDb(tables.executionQueue.select().orderBy("id", "desc"))
   ) as any[];
 
-  // Get running executions with SQL filtering for better performance
-  const runningExecutions = store.useQuery(
-    queryDb(
-      tables.executionQueue
-        .select()
-        .where({
-          status: { op: "IN", value: ["executing", "pending", "assigned"] },
-        })
-        .orderBy("id", "desc")
-    )
-  ) as any[];
   const [showRuntimeHelper, setShowRuntimeHelper] = React.useState(false);
   const [focusedCellId, setFocusedCellId] = React.useState<string | null>(null);
   const [contextSelectionMode, setContextSelectionMode] = React.useState(false);
   const hasEverFocusedRef = React.useRef(false);
-
-  const currentNotebookId = getCurrentNotebookId();
-  const runtimeCommand = getRuntimeCommand(currentNotebookId);
-
-  // Check runtime status
-  const getRuntimeHealth = (session: RuntimeSessionData) => {
-    if (session.status === "starting") {
-      // If session is starting, it's connecting
-      return session.isActive ? "connecting" : "unknown";
-    }
-    if (!session.isActive) {
-      return "disconnected";
-    }
-    // For active sessions, use status to determine health
-    switch (session.status) {
-      case "ready":
-      case "busy":
-        return "healthy";
-      case "restarting":
-        return "warning";
-      case "terminated":
-        return "disconnected";
-      default:
-        return "unknown";
-    }
-  };
-
-  const activeRuntime = runtimeSessions.find(
-    (session: RuntimeSessionData) =>
-      session.status === "ready" || session.status === "busy"
-  );
-  const hasActiveRuntime = Boolean(
-    activeRuntime &&
-      ["healthy", "warning", "connecting"].includes(
-        getRuntimeHealth(activeRuntime)
-      )
-  );
-  const runtimeHealth = activeRuntime
-    ? getRuntimeHealth(activeRuntime)
-    : "disconnected";
-  const runtimeStatus =
-    activeRuntime?.status ||
-    (runtimeSessions.length > 0 ? runtimeSessions[0].status : "disconnected");
-
-  const copyRuntimeCommand = useCallback(() => {
-    navigator.clipboard.writeText(runtimeCommand);
-    // Could add a toast notification here
-  }, [runtimeCommand]);
-
-  const interruptAllExecutions = useCallback(async () => {
-    // Cancel each running execution (already filtered by SQL query)
-    for (const execution of runningExecutions) {
-      store.commit(
-        events.executionCancelled({
-          queueId: execution.id,
-          cellId: execution.cellId,
-          cancelledBy: currentUserId,
-          reason: "User interrupted all executions from runtime UI",
-        })
-      );
-    }
-  }, [runningExecutions, store, currentUserId]);
 
   // Prefetch output components adaptively based on connection speed
   React.useEffect(() => {
@@ -501,21 +429,7 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
                       {metadata.find((m) => m.key === "runtimeType")?.value ??
                         "python3"}
                     </span>
-                    <Circle
-                      className={`h-2 w-2 fill-current ${
-                        activeRuntime && runtimeHealth === "healthy"
-                          ? "text-green-500"
-                          : activeRuntime && runtimeHealth === "warning"
-                            ? "text-amber-500"
-                            : activeRuntime && runtimeHealth === "connecting"
-                              ? "text-blue-500"
-                              : activeRuntime && runtimeHealth === "warning"
-                                ? "text-amber-500"
-                                : runtimeStatus === "starting"
-                                  ? "text-blue-500"
-                                  : "text-red-500"
-                      }`}
-                    />
+                    <RuntimeHealthIndicator />
                   </Button>
                   <Button
                     variant={contextSelectionMode ? "default" : "outline"}
@@ -538,231 +452,11 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
               </div>
             </div>
 
-            {showRuntimeHelper && (
-              <div className="bg-card border-t">
-                <div className="w-full px-3 py-4 sm:mx-auto sm:max-w-6xl sm:px-4">
-                  <div className="mb-3 flex items-center justify-between">
-                    <h4 className="flex items-center gap-2 text-sm font-medium">
-                      Runtime Status
-                      <Circle
-                        className={`h-2 w-2 fill-current ${
-                          activeRuntime && runtimeHealth === "healthy"
-                            ? "text-green-500"
-                            : activeRuntime && runtimeHealth === "warning"
-                              ? "text-amber-500"
-                              : activeRuntime && runtimeHealth === "connecting"
-                                ? "text-blue-500"
-                                : activeRuntime && runtimeHealth === "warning"
-                                  ? "text-amber-500"
-                                  : runtimeStatus === "starting"
-                                    ? "text-blue-500"
-                                    : "text-red-500"
-                        }`}
-                      />
-                      <span
-                        className={`text-xs ${
-                          activeRuntime && runtimeHealth === "healthy"
-                            ? "text-green-600"
-                            : activeRuntime && runtimeHealth === "warning"
-                              ? "text-amber-600"
-                              : activeRuntime && runtimeHealth === "connecting"
-                                ? "text-blue-600"
-                                : activeRuntime && runtimeHealth === "warning"
-                                  ? "text-amber-600"
-                                  : runtimeStatus === "starting"
-                                    ? "text-blue-600"
-                                    : "text-red-600"
-                        }`}
-                      >
-                        {activeRuntime && runtimeHealth === "healthy"
-                          ? "Connected"
-                          : activeRuntime && runtimeHealth === "warning"
-                            ? "Connected (Slow)"
-                            : activeRuntime && runtimeHealth === "connecting"
-                              ? "Connecting..."
-                              : activeRuntime && runtimeHealth === "warning"
-                                ? "Connected (Warning)"
-                                : runtimeStatus === "starting"
-                                  ? "Starting"
-                                  : "Disconnected"}
-                      </span>
-                    </h4>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      onClick={() => setShowRuntimeHelper(false)}
-                      className="h-6 w-6 p-0"
-                    >
-                      ×
-                    </Button>
-                  </div>
-
-                  {!hasActiveRuntime && (
-                    <>
-                      <p className="text-muted-foreground mb-3 text-sm">
-                        Run this command in your terminal to start a runtime for
-                        notebook{" "}
-                        <code className="bg-muted rounded px-1">
-                          {currentNotebookId}
-                        </code>
-                        :
-                      </p>
-                      <div className="flex items-center gap-2 rounded bg-slate-900 p-3 font-mono text-sm text-slate-100">
-                        <span className="flex-1">{runtimeCommand}</span>
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          onClick={copyRuntimeCommand}
-                          className="h-8 w-8 p-0 text-slate-300 hover:bg-slate-700 hover:text-slate-100"
-                        >
-                          <Copy className="h-4 w-4" />
-                        </Button>
-                      </div>
-                      <p className="text-muted-foreground mt-2 text-xs">
-                        Note: Each notebook requires its own runtime instance.
-                        The runtime will connect automatically once started.
-                      </p>
-                    </>
-                  )}
-
-                  {hasActiveRuntime && activeRuntime && (
-                    <div className="space-y-2 text-sm">
-                      <div className="flex justify-between">
-                        <span className="text-muted-foreground">
-                          Session ID:
-                        </span>
-                        <code className="bg-muted rounded px-1 text-xs">
-                          {activeRuntime.sessionId}
-                        </code>
-                      </div>
-                      <div className="flex justify-between">
-                        <span className="text-muted-foreground">
-                          Runtime Type:
-                        </span>
-                        <span>{activeRuntime.runtimeType}</span>
-                      </div>
-                      <div className="flex justify-between">
-                        <span className="text-muted-foreground">Status:</span>
-                        <span
-                          className={`font-medium ${
-                            activeRuntime.status === "ready"
-                              ? "text-green-600"
-                              : activeRuntime.status === "busy"
-                                ? "text-amber-600"
-                                : "text-red-600"
-                          }`}
-                        >
-                          {activeRuntime.status === "ready"
-                            ? "Ready"
-                            : activeRuntime.status === "busy"
-                              ? "Busy"
-                              : activeRuntime.status.charAt(0).toUpperCase() +
-                                activeRuntime.status.slice(1)}
-                        </span>
-                      </div>
-                      {activeRuntime.status && (
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">
-                            Last Heartbeat:
-                          </span>
-                          <span className="flex items-center gap-1 text-xs">
-                            Status: {activeRuntime.status}
-                          </span>
-                        </div>
-                      )}
-                      <div className="flex justify-between">
-                        <span className="text-muted-foreground">
-                          Capabilities:
-                        </span>
-                        <div className="flex gap-1">
-                          {activeRuntime.canExecuteCode && (
-                            <span className="rounded bg-blue-100 px-1 text-xs text-blue-800">
-                              Code
-                            </span>
-                          )}
-                          {activeRuntime.canExecuteSql && (
-                            <span className="rounded bg-purple-100 px-1 text-xs text-purple-800">
-                              SQL
-                            </span>
-                          )}
-                          {activeRuntime.canExecuteAi && (
-                            <span className="rounded bg-green-100 px-1 text-xs text-green-800">
-                              AI
-                            </span>
-                          )}
-                        </div>
-                      </div>
-
-                      {/* Interrupt Button */}
-                      {executionQueue.some(
-                        (exec: any) =>
-                          exec.status === "executing" ||
-                          exec.status === "pending" ||
-                          exec.status === "assigned"
-                      ) && (
-                        <div className="mt-4 border-t pt-4">
-                          <div className="flex items-center justify-between">
-                            <span className="text-muted-foreground text-sm">
-                              Running Executions: {runningExecutions.length}
-                            </span>
-                            <Button
-                              variant="destructive"
-                              size="sm"
-                              onClick={interruptAllExecutions}
-                              className="flex items-center gap-1"
-                            >
-                              <Square className="h-3 w-3" />
-                              <span>Interrupt All</span>
-                            </Button>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  {/* Show all runtime sessions for debugging */}
-                  {runtimeSessions.length > 1 && (
-                    <div className="mt-4 border-t pt-4">
-                      <h5 className="text-muted-foreground mb-2 text-xs font-medium">
-                        All Sessions:
-                      </h5>
-                      <div className="space-y-1">
-                        {runtimeSessions.map((session: RuntimeSessionData) => (
-                          <div
-                            key={session.sessionId}
-                            className="flex items-center justify-between text-xs"
-                          >
-                            <code className="bg-muted rounded px-1">
-                              {session.sessionId.slice(-8)}
-                            </code>
-                            <div className="flex items-center gap-2">
-                              <span
-                                className={`rounded px-1 ${
-                                  session.status === "ready"
-                                    ? "bg-green-100 text-green-800"
-                                    : session.status === "busy"
-                                      ? "bg-amber-100 text-amber-800"
-                                      : session.status === "terminated"
-                                        ? "bg-red-100 text-red-800"
-                                        : "bg-gray-100 text-gray-800"
-                                }`}
-                              >
-                                {session.status}
-                              </span>
-                              {session.status && (
-                                <span className="text-muted-foreground">
-                                  Status: {session.status}
-                                </span>
-                              )}
-                            </div>
-                          </div>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </div>
-            )}
+            <RuntimeHelper
+              showRuntimeHelper={showRuntimeHelper}
+              onClose={() => setShowRuntimeHelper(false)}
+              runtimeSessions={runtimeSessions as any[]}
+            />
           </div>
 
           <div
@@ -897,7 +591,7 @@ export const NotebookViewer: React.FC<NotebookViewerProps> = ({
                 cells={cells}
                 allRuntimeSessions={allRuntimeSessions}
                 executionQueue={executionQueue}
-                currentNotebookId={currentNotebookId}
+                currentNotebookId={getCurrentNotebookId()}
                 runtimeHealth={runtimeHealth}
               />
             </ErrorBoundary>
